Rename module contrast.js
